+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added folding ranges for multi-line function definitions, so long parameter lists can be collapsed
 - Added notification when we detect definitions file changes
 - Added support for auto-requiring modules when autocompleteing a type reference, to allow indexed types: i.e. autocompleting `Module` in `type Foo = Module.Bar`
+- Added `luau-lsp.require.directoryAliases` to map require string _prefixes_ to directories
+- Added tilde expansion to `luau-lsp.require.fileAliases` (and `directoryAliases`), mapping `~/foo.lua` to a file in your home directory
 - End autocompletion will now work for defined anonymous functions inside of function calls
 
 ### Changed