+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.21.0] - 2023-06-14
+
 ### Deprecated
 
 - Deprecated `luau-lsp.autocompleteEnd` setting in favour of `luau-lsp.completion.autocompleteEnd`